# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402 Bazaar API",
    "description": "The decentralized API marketplace where AI agents pay with USDC via HTTP 402 protocol. Free endpoints return data directly. Paid endpoints return HTTP 402 with payment details (wallet address, amount in USDC, supported chains). To use paid endpoints, send USDC on Base to the specified wallet, then retry with X-Payment-TxHash header. 69+ total endpoints available — this schema covers the 30 most popular ones.",
    "version": "2.0.0",
    "contact": {
      "name": "x402 Bazaar",
      "url": "https://x402bazaar.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/Wintyx57/x402-backend/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://x402-api.onrender.com",
      "description": "Production (Base mainnet)"
    }
  ],
  "paths": {
...
    "/api/agent/{agentId}": {
      "get": {
        "operationId": "lookupAgent",
        "summary": "ERC-8004 agent identity lookup",
        "description": "Look up an AI agent's on-chain identity via ERC-8004 registry on Base. Returns agent name, description, owner, and reputation. Free endpoint.",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "description": "The numeric agent ID in the ERC-8004 registry",
            "schema": { "type": "string", "pattern": "^\\d+$" }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent information",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentInfo" } } }
          },
          "404": { "description": "Agent not found" }
        }
      }
    },
...
    "/api/joke": {
      "get": {
        "operationId": "getRandomJoke",
        "summary": "Random joke (0.01 USDC)",
        "description": "Returns a random joke (setup + punchline or single-liner). Costs 0.01 USDC.",
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "A random joke", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
...
    "/api/quote": {
      "get": {
        "operationId": "getRandomQuote",
        "summary": "Random advice quote (0.005 USDC)",
        "description": "Returns a random piece of advice or inspirational quote. Costs 0.005 USDC.",
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "A random quote", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/facts": {
      "get": {
        "operationId": "getRandomFact",
        "summary": "Random fun fact (0.005 USDC)",
        "description": "Returns a random interesting fact. Costs 0.005 USDC.",
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "A random fact", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
...
    "/api/code": {
      "post": {
        "operationId": "executeCode",
        "summary": "Execute code in 50+ languages (0.005 USDC)",
        "description": "Execute code in 50+ programming languages via a sandboxed environment. Costs 0.005 USDC.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["language", "code"],
                "properties": {
                  "language": { "type": "string", "description": "Programming language (e.g., python, javascript, go, rust)" },
                  "code": { "type": "string", "description": "Source code to execute" }
                }
              }
            }
          }
        },
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "Code execution result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentInfo": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "agentId": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "owner": { "type": "string" },
          "reputation": { "type": "integer" }
        }
      },
      "PaymentRequired": {
        "type": "object",
        "description": "Returned when payment is required. Contains details for making a USDC payment on Base blockchain.",
        "properties": {
          "error": { "type": "string", "example": "Payment Required" },
          "x402": {
            "type": "object",
            "properties": {
              "version": { "type": "string", "example": "1" },
              "payload": {
                "type": "object",
                "properties": {
                  "amount": { "type": "string", "description": "Amount in USDC micro-units (6 decimals)", "example": "5000" },
                  "amountHuman": { "type": "string", "description": "Human-readable USDC amount", "example": "0.005 USDC" },
                  "recipient": { "type": "string", "description": "Wallet address to send USDC to", "example": "0xfb1c478BD5567BdcD39782E0D6D23418bFda2430" },
                  "chain": { "type": "string", "example": "base" },
                  "chainId": { "type": "integer", "example": 8453 },
                  "token": { "type": "string", "example": "USDC" },
                  "tokenAddress": { "type": "string", "example": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913" }
                }
              }
            }
          },
          "instructions": { "type": "string", "description": "How to complete payment and retry the request" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        },
        "description": "Response varies by endpoint. Always includes success:true and endpoint-specific data."
      }
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. Send USDC on Base to the specified wallet address, then retry with X-Payment-TxHash header.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/PaymentRequired" }
          }
        }
      }
    }
  }
}